rm: old files, duplicates
