+++ lib/path/__init__.py
@@ -1,0 +1,8 @@
+PATH_TYPES = ['first', 'best']
+MAP_NAME_KEY = ['map_name']
+PATH_METRICS_KEYS = ['cost', 'time_sec', 'time_it', 'samples_taken', 'nodes_taken']
+RUNS_METRICS = ['costs', 'paths', 'samples', 'nodes']
+RUNS_COUNTS = ['mean', 'std']
+
+
+
