--- results/index.html
@@ -1,18 +1,0 @@
-<html>
-    <head>
-    </head>
-    <body>
-        <h1>Averaged stats demo for GAN and Pix2Pix</h1>
-        <h3>(see pages below for details)</h3>
-        <div>
-            <ul>
-            <li><p><a href="box_plots.html">Box plots with averaged stats</a></p>
-            <li><p><a href="line_plots.html">Line plots (include std)</a></p> 
-            </ul>
-        </div>
-        <div>
-            <img src="collected_stats_gan_best_cost.png">
-        </div>
-    </body>
-
-</html>
